twitter-producer-svc/requirements.txt
kafka-python
python-twitter
tweepy
#confluent-kafka